update generate command

+++ src/cli/commands/generate.ts
@@ -1,4 +1,4 @@
-import { outputFile } from 'fs-extra';
+import { outputFile, pathExists } from 'fs-extra';
 import chalk from 'chalk';
 
 import { GenerateService, GenerateType } from '../../public-api';
@@ -9,12 +9,16 @@ export class GenerateCommand {
 
   async run(type: GenerateType, dest: string) {
     const { content, path, fullPath } = this.generateService.generate(type, dest);
-    // save the file
-    await outputFile(fullPath, content);
-    // modify
-    await this.generateService.modify(type, path);
-    // show result
-    return console.log(`New ${chalk.yellow(type)} saved: `, chalk.green(path));
+    if (await pathExists(fullPath)) {
+      return console.log(`A ${chalk.yellow(type)} already available at ` + chalk.green(path));
+    } else {
+      // save the file
+      await outputFile(fullPath, content);
+      // modify
+      await this.generateService.modify(type, path);
+      // show result
+      return console.log(`New ${chalk.yellow(type)} saved at ` + chalk.green(path));
+    }
   }
 
 }
